QR Purchase email, final logic details

+++ src/main/java/com/uq/jokievents/model/OrderPayment.java
@@ -8,6 +8,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -25,5 +26,5 @@ public class OrderPayment {
     private String paymentType;
     private String paymentCurrency;
     private String authorizationCode;
-    private float paymentValue;
+    private BigDecimal paymentValue;
 }
